requirements.txt
streamlit>=1.32.0
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
ta>=0.11.0
plotly>=5.18.0
requests>=2.31.0
streamlit-autorefresh>=1.0.1
scipy>=1.11.0
finnhub-python>=2.4.20
